% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-singapore.R
\docType{data}
\name{buildings_pop_sgp}
\alias{buildings_pop_sgp}
\title{Population count per residential building in Singapore}
\format{
\code{sf} polygons.
}
\source{
Building polygons \href{https://www.openstreetmap.org/copyright}{copyrighted} OpenStreetMap contributors and available from https://www.openstreetmap.org.
Made available via the \href{https://opendatacommons.org/licenses/odbl/summary/}{ODbL License}.

Population data and census block polygons \code{data(pop_sgp)} are from the \href{https://www.singstat.gov.sg/find-data/search-by-theme/population/geographic-distribution/latest-data}{Department of Statistics Singapore};
and \href{https://data.gov.sg/dataset/master-plan-2019-subzone-boundary-no-sea}{Singapore Master Plan Subzones}, respectively;
land use polygons \code{data(landuse_sgp)} are from the Singapore Land Use Master Plan released in \href{https://data.gov.sg/dataset/master-plan-2019-land-use-layer}{2019}.
All are made available under the terms of the \href{https://data.gov.sg/open-data-licence}{Singapore Open Data Licence version 1.0}.
}
\usage{
buildings_pop_sgp
}
\description{
Example (random 5\% subset) dataset of residential building polygons in Singapore,
each with a population count (column \code{popcount})
estimated via dasymetric mapping.
}
\details{
Building polygons were downloaded from OpenStreetMap
(data snapshot on \code{2021-01-01} from the \href{https://download.geofabrik.de}{Geofabrik database}),
using the function \code{get_buildings_osm()}. The population count per census block in the year 2020
was re-distributed across the buildings located within residential land use zones,
by performing dasymetric mapping using the functions \code{pop_dasymap()} and \code{pop_density_polygonise()}.
See vignette and examples in \code{pop_density_polygonise()} for more details.
The dataset is a random 5\% subset of the resulting polygons.
}
\examples{
data(buildings_pop_sgp)
head(buildings_pop_sgp)
}
\keyword{datasets}
