% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-singapore.R
\docType{data}
\name{buildings_sgp}
\alias{buildings_sgp}
\title{Buildings in Singapore}
\format{
\code{sf} polygons. The following columns are used in this package:
\describe{
\item{levels}{Number of building levels, derived from \code{building_levels}.
Values were set to \code{1} if the extracted value was empty or \code{NA}, and set to \code{NA} if \verb{≤ 0} (i.e. underground);
values were then rounded up to the nearest whole number.}
}
}
\source{
Map data \href{https://www.openstreetmap.org/copyright}{copyrighted} OpenStreetMap contributors and available from https://www.openstreetmap.org.
}
\usage{
buildings_sgp
}
\description{
Example dataset of building polygons in Singapore downloaded from OpenStreetMap
(data snapshot on \code{2021-01-01} from the from the \href{https://download.geofabrik.de}{Geofabrik database}),
using the function \code{get_buildings_osm()}.
}
\examples{
data(buildings_sgp)
head(buildings_sgp)
}
\keyword{datasets}
