% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_buildings_osm.R
\name{get_buildings_osm}
\alias{get_buildings_osm}
\title{Get building polygons from OpenStreetMap}
\usage{
get_buildings_osm(
  place,
  date = NULL,
  dir_raw = oe_download_directory(),
  filename = NULL,
  ...
)
}
\arguments{
\item{place}{\code{sf} object (with projected coordinate reference system). Geographical area to match with the (\code{.osm.pbf}) file in the data archive.
Argument passed to \code{osmextract::oe_match()}.}

\item{date}{Date of OSM data snapshot to download. Refer to https://download.geofabrik.de
for the specific dates available. Defaults to \code{NULL} (download the latest available data).}

\item{dir_raw}{character. Directory to download the raw unprocessed OSM data. Passed to
argument \code{download_directory} in \code{osmextract::oe_read()}.}

\item{filename}{character (optional). File path to export output data (GeoJSON format).}

\item{...}{Other arguments passed to \code{osmextract::oe_read()}.}
}
\value{
The processed building polygons (\code{sf} object).
}
\description{
Download and process OpenStreetMap (OSM) building polygons within a specified geographical \code{place},
from the \href{https://download.geofabrik.de}{Geofabrik database}. It is a wrapper around
functions in the package \href{https://docs.ropensci.org/osmextract/index.html}{\code{osmextract}}, and
processes the downloaded files for subsequent analyses. Refer to package \code{osmextract} for
more details and options for input arguments when downloading the data.
}
\details{
Data is filtered by key-value attributes, where \verb{building:} is not \code{NULL}.
The column \code{levels} is derived from \code{building:levels}; values were set to \code{1} if the
extracted value is empty or \code{NA}, and set to \code{NA} if \verb{≤ 0} (i.e. underground);
values were then rounded up to the nearest whole number.
The column \code{area_m2} represents the building footprint area,
and \code{floorarea_m2} is calculated by multiplying the \code{area_m2} by the number of \code{levels}.
}
