% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_amenities_osm.R
\name{get_playgrounds_osm}
\alias{get_playgrounds_osm}
\title{Get public playgrounds from OpenStreetMap}
\usage{
get_playgrounds_osm(
  place,
  date = NULL,
  dir_raw = oe_download_directory(),
  filename = NULL,
  ...
)
}
\arguments{
\item{place}{\code{sf} object (with projected coordinate reference system). Geographical area to match with the (\code{.osm.pbf}) file in the data archive.
Argument passed to \code{osmextract::oe_match()}.}

\item{date}{Date of OSM data snapshot to download. Refer to https://download.geofabrik.de
for the specific dates available. Defaults to \code{NULL} (download the latest available data).}

\item{dir_raw}{character. Directory to download the raw unprocessed OSM data. Passed to
argument \code{download_directory} in \code{osmextract::oe_read()}.}

\item{filename}{character (optional). File path to export output data (GeoJSON format).}

\item{...}{Other arguments passed to \code{osmextract::oe_read()}.}
}
\value{
The processed playgrounds (\code{sf} object).
}
\description{
Download and process OpenStreetMap (OSM) public playgrounds (points) within a specified geographical \code{place},
from the \href{https://download.geofabrik.de}{Geofabrik database}. It is a wrapper around
functions in the package \href{https://docs.ropensci.org/osmextract/index.html}{\code{osmextract}}, and
processes the downloaded files for subsequent analyses. Refer to package \code{osmextract} for
more details and options for input arguments when downloading the data.
}
\details{
OSM points filtered by key-value attributes, where \code{leisure:fitness_station} or \verb{sport:*},
and \verb{access:} is not \code{no} or \code{private}.
}
\examples{
\dontrun{
city_boundaries <- data(singapore) \%>\%
   dplyr::filter(year == 2020) \%>\%
   sf::st_union() \%>\%
   sf::st_as_sf() \%>\%
   smoothr::fill_holes(threshold = units::set_units(1, "km^2"))  \%>\%
   smoothr::drop_crumbs(threshold = units::set_units(1, "km^2"))  \%>\%
   sf::st_make_valid()

get_playgrounds_osm(place = city_boundaries,
                    date = as.Date("2021-01-01"),
                    filename = "public-playgrounds_osm-points_2021-01-01.geojson")
}

}
