% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-singapore.R
\docType{data}
\name{landuse_sgp}
\alias{landuse_sgp}
\title{Land use master plan for the city of Singapore}
\format{
\code{sf} polygons of land use released in years 2014 and 2019.
Data is in the 'long' format (rows are repeated for each year).
The following columns are used in this package:
\describe{
\item{lu_desc}{Land use type (description)}
\item{year}{Year of Master Plan}
}
}
\source{
Polygons from the Singapore Land Use Master Plans released in the years
\href{https://data.gov.sg/dataset/master-plan-2014-land-use}{2014} and
\href{https://data.gov.sg/dataset/master-plan-2019-land-use-layer}{2019}.
Made available under the terms of the \href{https://data.gov.sg/open-data-licence}{Singapore Open Data Licence version 1.0}.
}
\usage{
landuse_sgp
}
\description{
Example dataset containing Singapore Master Plan Land Use Zones for the years
\href{https://data.gov.sg/dataset/master-plan-2014-land-use}{2014} and
\href{https://data.gov.sg/dataset/master-plan-2019-land-use-layer}{2019}.
Released under the \href{https://data.gov.sg/open-data-licence}{Singapore Open Data License}.
}
\examples{
data(landuse_sgp)
head(landuse_sgp)
}
\keyword{datasets}
