% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchyear.R
\name{matchyear}
\alias{matchyear}
\title{Match the year between the dataset of interest and another data source}
\usage{
matchyear(
  data,
  data_tomatch,
  match = c("exact", "closest", "recent", "soonest"),
  year = NULL
)
}
\arguments{
\item{data}{Tabular object (e.g. \code{data.frame}, \code{sf}, \code{tibble}) containing the data.}

\item{data_tomatch}{Tabular object (e.g. \code{data.frame}, \code{sf}, \code{tibble}) to match the data to.}

\item{match}{Type of matching; either \code{'exact'}, \code{'closest'}, \code{'recent'} or \code{'soonest'}.}

\item{year}{Specify column name for the year within both datasets.
Columns in both datasets should be numeric.}
}
\value{
\code{data} with additional column \code{year_match},
representing the matching year of the other dataset.
}
\description{
Helper function to match the year in the data with either the (1) exact;
(2) most recent (past); (3) closest (past or future); or (4) soonest (future) year in another dataset.
The column name within both datasets that represent the year must be specified.
Dataset containing multiple years must be in the 'long' format.
}
