% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons_clean.R
\name{polygons_clean}
\alias{polygons_clean}
\title{Clean up polygons after download}
\usage{
polygons_clean(
  input,
  snap_tolerance = 5,
  min_area = units::set_units(0, "m^2"),
  aggregate_polygons = 15
)
}
\arguments{
\item{input}{\code{sf} object (with projected coordinate reference system).}

\item{snap_tolerance}{numeric. Argument for \code{tolerance} level passed to \code{sf::st_snap()},
used to rectify nearly coincident edges between polygons. Provided either as a units object (see \code{units::set_units()}),
or a number in the units of the coordinate reference system.
After \code{sf::st_snap()} is run, a polygon that is \href{https://postgis.net/docs/ST_Covers.html}{covered} by another will be removed.
Defaults to \code{5}. Set to \code{0} if you do not wish to rectify minor overlaps.}

\item{min_area}{numeric. Specify minimum area of each polygon to be retained in the output,
passed to argument \code{threshold} in \code{smoothr::drop_crumbs()}.
Provided either as a units object (see \code{units::set_units()}), or a number in the units of
the coordinate reference system. Defaults to \code{0} \eqn{m^2}.}

\item{aggregate_polygons}{numeric. Argument for \code{dist} passed to \code{sf::st_buffer()}.
Buffered polygons that overlap will be aggregated into multipolygons.
Set to \code{NULL} if you do not wish to aggregate to multipolygons.}
}
\value{
The processed polygons (\code{sf} object).
}
\description{
Helper function to process and clean up polygons downloaded from OpenStreetMap.
}
\details{
Polygons \href{https://postgis.net/docs/ST_Contains.html}{contained} in or \href{https://postgis.net/docs/ST_Overlaps.html}{overlapping} others are removed.
Polygons vertices are then \href{https://postgis.net/docs/ST_Snap.html}{snapped}
at a tolerance level based on \code{snap_tolerance} (if specified) to rectify nearly coincident edges,
and polygons \href{https://postgis.net/docs/ST_Covers.html}{covered} by others are subsequently removed. Polygons are then aggregated
to multipolygons based on distance set in \code{aggregate_polygons} (if specified).
All spatial relations (italicised) follow the \href{https://en.wikipedia.org/wiki/DE-9IM}{DE-9IM} standard.
}
