% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons_mutually_exclude.R
\name{polygons_mutually_exclude}
\alias{polygons_mutually_exclude}
\title{Make polygon input mutually exclusive (no overlaps) with another}
\usage{
polygons_mutually_exclude(input, mutually_exclude, snap_tolerance = 0)
}
\arguments{
\item{input}{\code{sf} object to process (with projected coordinate reference system).}

\item{mutually_exclude}{\code{sf} object. \code{input} polygons \href{https://postgis.net/docs/ST_Contains.html}{contained}
within this object, and \href{https://postgis.net/docs/ST_Intersection.html}{intersections} between the two, will be removed.
Should have the same coordinate reference system as \code{input}.}

\item{snap_tolerance}{numeric. Argument for \code{tolerance} level passed to \code{sf::st_snap()},
used to rectify nearly coincident edges between polygons before running \code{sf::st_contains()}.
Provided either as a units object (see \code{units::set_units()}),
or a number in the units of the coordinate reference system. Defaults to \code{0}.}
}
\value{
The processed \code{input} polygons (\code{sf} object).
}
\description{
Helper function to ensure that polygons (e.g. parks, beaches, informal nature areas)
are mutually-exclusive (i.e. non-overlapping).
}
\details{
Polygons vertices first are \href{https://postgis.net/docs/ST_Snap.html}{snapped} at a tolerance level
based on \code{snap_tolerance} (if specified) to rectify nearly coincident edges,
before removing \code{input} polygons \href{https://postgis.net/docs/ST_Contains.html}{contained} in \code{mutually_exclude}.
\href{https://postgis.net/docs/ST_Intersection.html}{Intersections} between the two are then removed.
}
