% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_dasymap.R
\name{pop_dasymap}
\alias{pop_dasymap}
\title{Perform dasymetric mapping of population data for a single time period}
\usage{
pop_dasymap(
  pop_polygons,
  pop_perblock_count,
  pop_perblock_density,
  land_relative_density,
  filename = NULL,
  dir_rastertemplate = NULL,
  overwrite = TRUE,
  ...
)
}
\arguments{
\item{pop_polygons}{Simple feature (polygons) of the population data for a single time period.
May be extracted from the \code{pop_polygons} list element generated by \code{rasterise_pop()}.}

\item{pop_perblock_count}{Raster of population count per census block for a single time period.
May be extracted from the \code{pop_count_rasters} list element generated by \code{rasterise_pop()}.}

\item{pop_perblock_density}{Raster of population density per census block for a single time period.
May be extracted from the \code{pop_density_rasters} list element generated by \code{rasterise_pop()}.}

\item{land_relative_density}{Raster of the relative density (e.g. suitability, habitability) of land for a single time period.
May be extracted from results generated by \code{rasterise_buildings()} or \code{rasterise_landuse()}.}

\item{filename}{character (optional). Export output raster to disk.}

\item{dir_rastertemplate}{character. Filepath to the raster used to define the pixel resolution, extent, nrow, ncol of
the output raster; object is passed to the '\code{y}' argument in \code{terra::rasterize()}.
Defaults to the template raster generated by the function \code{rasterise_pop()} within the \code{tempdir()}.}

\item{overwrite}{logical. Argument passed to \code{terra::writeRaster()}. If \code{TRUE}, \code{filename} is overwritten.}

\item{...}{Other arguments passed to \code{terra::writeRaster()}.}
}
\value{
Raster of population density, based on relative density values of the land defined in \code{land_relative_density}.
}
\description{
Distribute population data per census block across inhabitable land as defined by \code{land_relative_density}.
Data provided should be a snapshot of the population for a single time period (e.g. one year).
All input data should have a similar projected coordinate reference system
specific to the target area.
}
