% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-singapore.R
\docType{data}
\name{pop_sgp}
\alias{pop_sgp}
\title{Population counts within census block polygons in Singapore}
\format{
\code{sf} polygons with census data for years 2018 and 2020.
Data is in the 'long' format (rows are repeated for each census year).
The following columns are used in this package:
\describe{
\item{subzone_n}{Census block name}
\item{year}{Census year}
\item{pop_count}{Population count within census block}
}
}
\source{
Contains census data from the \href{https://www.singstat.gov.sg/find-data/search-by-theme/population/geographic-distribution/latest-data}{Department of Statistics Singapore}
and polygons from the \href{https://data.gov.sg/dataset/master-plan-2019-subzone-boundary-no-sea}{Singapore Master Plan Subzones},
both of which are made available under the terms of the \href{https://data.gov.sg/open-data-licence}{Singapore Open Data Licence version 1.0}.
}
\usage{
pop_sgp
}
\description{
Example dataset containing Singapore census data for years 2018 and 2020 from the \href{https://www.singstat.gov.sg/find-data/search-by-theme/population/geographic-distribution/latest-data}{Department of Statistics Singapore},
joined by name to the Master Plan Subzone polygons from \href{https://data.gov.sg/dataset/master-plan-2019-subzone-boundary-no-sea}{data.gov.sg}.
Both datasets are released under the \href{https://data.gov.sg/open-data-licence}{Singapore Open Data License}.
}
\examples{
data(pop_sgp)
head(pop_sgp)
}
\keyword{datasets}
