% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parks_calc_attributes_helpers.R
\name{raster_class_area}
\alias{raster_class_area}
\title{Calculate the total class area of a (classified) raster for each park polygon.}
\usage{
raster_class_area(
  polygons,
  raster,
  raster_min_patch_size = units::set_units(0, "m^2"),
  relative = TRUE,
  ...
)
}
\arguments{
\item{polygons}{\code{sf} (with projected coordinate reference system).}

\item{raster}{\code{SpatRaster} object from \code{terra::rast()}. Should have a (projected) coordinate reference system similar to \code{polygons}.}

\item{raster_min_patch_size}{Minimum patch size to be included in results.
Provided either as a units object (see \code{units::set_units()}),
or a number in the units of the coordinate reference system. Defaults to \code{0} m^2.}

\item{relative}{logical. Whether or not to calculate relative amounts
(i.e. percentage area). Defaults to \code{TRUE}.}

\item{...}{Other arguments passed to \code{terra::extract()}.}
}
\value{
\code{polygons} with added column(s) \verb{< class value >_area}, and \verb{< class value >_area_pct} if \code{relative} is set to \code{TRUE}.
Note that the value \code{0} will be summarised; convert pixels to \code{NA} if you wish to exclude them.
}
\description{
Helper function within \code{parks_calc_attributes()}.
For each unique pixel value (each representing a specific raster class/category),
the total area per class is calculated.
}
