% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterise_buildings.R
\name{rasterise_buildings}
\alias{rasterise_buildings}
\title{Rasterise building polygons}
\usage{
rasterise_buildings(
  sf_buildings,
  proxy_pop_density = NULL,
  year = "year",
  sf_pop = NULL,
  sf_landuse = NULL,
  match_buildings_pop = "closest",
  match_buildings_landuse = "closest",
  dir_rastertemplate = NULL,
  dir_processing = tempdir(),
  dir_export = NULL,
  overwrite = TRUE,
  wopt = list(gdal = c("COMPRESS=LZW")),
  ...
)
}
\arguments{
\item{sf_buildings}{\code{sf }polygons of the buildings.
Data should be in a projected coordinate reference system.}

\item{proxy_pop_density}{character. Specify column name of the building attribute that is used
as an indicator of population density (e.g. building height, no. of levels).
This column will be used to assign values to the output raster. If not provided, population density
across all building pixels are assumed to be similar.}

\item{year}{character. Specify column name for the year within \code{sf_buildings}, \code{sf_landuse}
and \code{sf_pop} (if provided). Defaults to \code{'year'}. Column data should be numeric.}

\item{sf_pop}{(optional) \code{sf} polygons containing the population census data with column containing the census year.
If absent, the output (building rasters) will not be associated with specific population census year(s).}

\item{sf_landuse}{(optional) \code{sf} polygons of the land use zones.}

\item{match_buildings_pop}{character. Type of matching between \code{sf_buildings} and \code{sf_pop} (if provided), passed on to \code{match} argument in function \code{matchyear()}.
Either \code{'exact'}, \code{'closest'}, \code{'recent'} or \code{'soonest'}. Defaults to \code{'closest'}.}

\item{match_buildings_landuse}{character. Type of matching between \code{sf_buildings} and \code{sf_landuse} (if provided), passed on to \code{match} argument in function \code{matchyear()}.
Either \code{'exact'}, \code{'closest'}, \code{'recent'} or \code{'soonest'}. Defaults to \code{'closest'}.
This argument is used only if \code{sf_pop} is not provided.}

\item{dir_rastertemplate}{character. Filepath to the raster used to define the pixel resolution, extent, nrow, ncol of
the output raster; object is passed to the '\code{y}' argument in \code{terra::rasterize()}.
Defaults to the template raster generated by the function \code{rasterise_pop()} within \code{dir_processing} (see next argument).}

\item{dir_processing}{character. Directory to get intermediate files generated in previous steps (e.g. rasterised population/land use data).
Defaults to \code{tempdir()}.}

\item{dir_export}{character (optional). File path to directory to export output raster(s) to.}

\item{overwrite}{logical. Argument passed to \code{terra::writeRaster()}. Defaults to \code{TRUE}.}

\item{wopt}{list. Argument passed to \code{terra::writeRaster()}.}

\item{...}{Other arguments passed to \code{terra::writeRaster()}.}
}
\value{
List of raster files. Zero values are converted to \code{NA}.
}
\description{
Rasterise building polygons (\code{sf_buildings}) with reference to population (\code{sf_pop}) and/or
land use data (\code{sf_landuse}), if supplied. Multiple output rasters will be generated if
building data for multiple years are present.
}
\details{
If population (\code{sf_pop}) and/or land use data (\code{sf_landuse}) are supplied,
their raster(s) must have been previously generated in the \code{tempdir()} using
the functions \code{rasterise_pop()} and \code{rasterise_landuse()}, respectively.
Rasterised buildings will be masked away (i.e., convert pixels to \code{NA})
at areas with no (zero) population and/or land use data, according to the respective matching year(s).
Custom ways to match the years between datasets can be set via the helper function \code{matchyear()}.
The argument \code{match_buildings_pop} provides ways to match each year in the \code{sf_pop} (if provided)
to a specific year in \code{sf_buildings}. Building raster(s) will then be associated
with a specific population census year, and removed if there are no matching population census years.

If necessary, the argument \code{match_buildings_landuse} provides ways to match each year in \code{sf_buildings}
with a specific year in \code{sf_landuse}, but only if \code{sf_pop} is not provided. This is because the main
reference point for all the datasets is \code{sf_pop} (primary focus is the analysis of population data).
}
