% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterise_landuse.R
\name{rasterise_landuse}
\alias{rasterise_landuse}
\title{Rasterise residential land use polygons}
\usage{
rasterise_landuse(
  sf_landuse,
  subset = NULL,
  land_use = NULL,
  year = "year",
  sf_pop = NULL,
  match_landuse_pop = "recent",
  dir_rastertemplate = NULL,
  dir_processing = tempdir(),
  dir_export = tempdir(),
  overwrite = TRUE,
  wopt = list(gdal = c("COMPRESS=LZW")),
  ...
)
}
\arguments{
\item{sf_landuse}{Simple feature (polygons) of the land use zones.
Should be transformed to a projected coordinate reference system.}

\item{subset}{Named vector containing the residential land use classes of interest to retain in column \code{land_use}.
Each element is named according to the integer that will be used
to represent this class in the output raster. Defaults to \code{NULL} (integers will be alphabetically
assigned to land use classes).}

\item{land_use}{Specify column name of the land use classes.}

\item{year}{Specify column name for the year within \code{sf_landuse}
and \code{sf_pop} (if provided). Defaults to \code{'year'}. Columns should be numeric.}

\item{sf_pop}{(optional) Simple feature (polygons) containing the population census data with column containing the census year (optional).
If absent, the output (land use rasters) will not be associated with specific population census years.}

\item{match_landuse_pop}{Type of matching between \code{sf_landuse} and \code{sf_pop}, passed on to \code{match} argument in function \code{matchyear()}.
Either \code{'exact'}, \code{'closest'}, \code{'recent'} or \code{'soonest'}. Defaults to \code{'recent'}, i.e.
assumes that land use is pre-planned (and followed by population changes), rather than a result of post-monitoring (e.g. remotely-sensed).}

\item{dir_rastertemplate}{character. Filepath to the raster used to define the pixel resolution, extent, nrow, ncol of
the output  raster; object is passed to the '\code{y}' argument in \code{terra::rasterize()}.
Defaults to the template raster generated by the function \code{rasterise_pop()} within \code{dir_processing} (see next argument).}

\item{dir_processing}{character. Directory to get intermediate files generated in previous steps (e.g. rasterised population data).
Defaults to \code{tempdir()}.}

\item{dir_export}{character. File path to directory to export output raster(s) to. Defaults to \code{tempdir()}.
Set to \code{NULL} if you do not wish to export output for subsequent processing.}

\item{overwrite}{logical. Argument passed to \code{terra::writeRaster()}. Defaults to \code{TRUE}.}

\item{wopt}{list. Argument passed to \code{terra::writeRaster()}.}

\item{...}{Other arguments passed to \code{terra::writeRaster()}.}
}
\value{
List of raster files. Zero values are converted to \code{NA}.
Each raster is also an intermediate file exported to the directory as defined by \code{dir_export}
(defaults to \code{tempdir()}) for further processing.
}
\description{
Rasterise land use zones (\code{sf_landuse}) with reference to population data (\code{sf_pop}), if supplied.
Multiple output rasters will be generated if land use data for multiple years are present.
If the land use data includes unnecessary land use classes, the argument \code{subset} and \code{landuse}
allows the user to first subset the data to polygons defined as 'residential' land use.
}
\details{
If population data (\code{sf_pop}) is supplied, its raster(s) must have been previously generated in the \code{tempdir()} using
the function \code{rasterise_pop()}. Each year in the population data is matched to a specific year in the land use data
(using the helper function \code{matchyear()}). Land use raster(s) will then
be associated with a specific population census year, and removed if there are no matching population census years.
Rasterised land use zones will be masked away (i.e., convert pixels to \code{NA})
at areas with no (zero) population data (for the respective matching year).
}
