% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterise_pop.R
\name{rasterise_pop}
\alias{rasterise_pop}
\title{Rasterise population counts within census block polygons}
\usage{
rasterise_pop(
  sf,
  res = 10,
  census_block = NULL,
  pop_count = NULL,
  year = "year",
  dir_processing = tempdir()
)
}
\arguments{
\item{sf}{Simple feature (polygons) containing the population census data.
Should be transformed to a projected coordinate reference system (for calculation of pixel resolution).}

\item{res}{Specify the pixel resolution for rasterisation (in metres). Defaults to \code{10}.}

\item{census_block}{Specify column name of the census block unique identifier (e.g. name).}

\item{pop_count}{Specify column name of the population counts per census block.}

\item{year}{Specify column name of the census year. Defaults to \code{'year'}.}

\item{dir_processing}{character. Directory to store intermediate files. Defaults to \code{tempdir()}.
Set to \code{NULL} if you do not wish to export intermediate (temporary) files for subsequent processing.}
}
\value{
List containing (1) rasterised population count per census block,
(2) rasterised population density per census block, and (3) processed polygons used to generate
these rasters. Named \code{pop_count_rasters}, \code{pop_density_rasters} and \code{pop_polygons}, respectively. List is nested if multiple years are present (one sub-list for each census year).
Zero values for population counts/density are converted to \code{NA}. Intermediate (raster) files are exported to \code{tempdir()} for further processing.
}
\description{
Convert population counts per census block (polygons)
into population density grid (raster) for subsequent analyses.
Each polygon's count is divided by its area, such that the the integrated
density over each census block should equal the original count.
Census data for multiple years can be processed (one raster per year).
The column name for the \code{year} must be specified, even if data does not contain multiple years
(combined data across multiple years must be in the 'long' format).
}
