% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_recre_perbuilding.R
\name{ss_recre_perbuilding}
\alias{ss_recre_perbuilding}
\title{Calculate cumulative supply of a park attribute per building polygon}
\usage{
ss_recre_perbuilding(park_attribute, dist_matrix, c = 1)
}
\arguments{
\item{park_attribute}{Vector of quantities of a specific attribute for all parks.
Length of vector is equal to the number of parks considered.}

\item{dist_matrix}{Matrix containing buildings (rows) and their pairwise distances to each park (columns).
Order of parks (columns) should be identical to the order of parks (elements) in \code{park_attribute}.}

\item{c}{Coefficient determining rate of decay in recreation supply with increasing distance}
}
\value{
A vector containing the cumulative supply values across all parks,
for each building (row) in the input matrix \code{dist_matrix}.
The length of the vector equals to the number of buildings considered.
}
\description{
The supply $S$ of the park attribute is calculated based on the following equation:
}
\details{
\enumerate{
\item \eqn{S = \sum\limits_{i=1}^{n} s_{i} \cdot e^{-cd_{i}}}
}

where

\eqn{S} = Total supply of a specific park attribute to the building from parks \eqn{i};
\eqn{i = 1,2,3}... \eqn{n}, \eqn{n} = total number of parks

\eqn{s_{i}} = Supply of a specific park attribute from park \eqn{i}.
A perfect positive linear association is assumed, since the focus is on supply metrics.
Other coefficients (e.g.\eqn{ks^{z}_{i}}) may be added to model (monotonically positive)
relationships with actual park use or downstream benefits.

\eqn{d_{i}} = Euclidean distance from the building to park \eqn{i}.

\eqn{c} = Coefficient determining rate of decay in supply \eqn{i} with increasing distance.
}
\examples{


}
