% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_buildings_osm.R
\name{get_buildings_osm}
\alias{get_buildings_osm}
\title{Get building polygons from OpenStreetMap}
\usage{
get_buildings_osm(
  place,
  date = NULL,
  dir_raw = osmextract::oe_download_directory(),
  filename = NULL,
  driver = "GeoJSON",
  delete_dsn = TRUE,
  append = NA,
  ...
)
}
\arguments{
\item{place}{\code{sf} object (with projected coordinate reference system). Geographical area to match with the (\code{.osm.pbf}) file in the data archive.
Argument passed to \code{osmextract::oe_match()}.}

\item{date}{Date of OSM data snapshot to download. Refer to https://download.geofabrik.de
for the specific dates available. Defaults to \code{NULL} (download the latest available data).}

\item{dir_raw}{character. Directory to download the raw unprocessed OSM data. Passed to
argument \code{download_directory} in \code{osmextract::oe_read()}.}

\item{filename}{character (optional). File path to export output data.}

\item{driver}{character (optional). Name of driver used to export output data, passed to \code{sf::st_write()}.
Defaults to "GeoJSON".}

\item{delete_dsn}{logical (optional). Passed to \code{sf::st_write()}.}

\item{append}{defaults to \code{NA}, which raises an error if a layer exists. Passed to \code{sf::st_write()}.}

\item{...}{Other arguments passed to \code{osmextract::oe_read()}.}
}
\value{
The processed building polygons (\code{sf} object).
}
\description{
Download and process OpenStreetMap (OSM) building polygons within a specified geographical \code{place},
from the \href{https://download.geofabrik.de}{Geofabrik database}. It is a wrapper around
functions in the package \href{https://docs.ropensci.org/osmextract/index.html}{\code{osmextract}}, and
processes the downloaded files for subsequent analyses. Refer to package \code{osmextract} for
more details and options for input arguments when downloading the data.
}
\details{
Data is filtered by key-value attributes, where \verb{building:} is not \code{NULL}.
The column \code{levels} is derived from \code{building:levels}; values were set to \code{1} if the
extracted value is empty or \code{NA}, and set to \code{NA} if \verb{≤ 0} (i.e. underground);
values were then rounded up to the nearest whole number.
The column \code{area_m2} represents the building footprint area,
and \code{floorarea_m2} is calculated by multiplying the \code{area_m2} by the number of \code{levels}.
}
\examples{
\dontrun{
data(pop_sgp)
pop_sgp <- sf::st_transform(pop_sgp, sf::st_crs(32648)) # transform to projected crs

# merge all census blocks for chosen year (2020) into single multi-polygon
# function requires that polygons are merged
city_boundaries <- pop_sgp \%>\%
   dplyr::filter(year == 2020) \%>\%
   sf::st_union() \%>\%
   sf::st_as_sf() \%>\%
   smoothr::fill_holes(threshold = units::set_units(1, 'km^2'))  \%>\% # clean up
   smoothr::drop_crumbs(threshold = units::set_units(1, 'km^2'))  \%>\%
   sf::st_make_valid()

# run function
get_buildings_osm(place = city_boundaries,
                  date = as.Date('2021-01-01'),
                  filename = 'buildings_osm-polygons_2021-01-01.geojson')
}

}
