% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-singapore.R
\docType{data}
\name{parks_sgp}
\alias{parks_sgp}
\title{Public parks in Singapore}
\format{
\code{sf} polygons.
}
\source{
Map data \href{https://www.openstreetmap.org/copyright}{copyrighted} OpenStreetMap contributors and available from https://www.openstreetmap.org.
}
\usage{
parks_sgp
}
\description{
Example dataset of park polygons in Singapore downloaded from OpenStreetMap
(data snapshot on \code{2021-01-01} from the from the \href{https://download.geofabrik.de}{Geofabrik database}),
using the function \code{get_parks_osm()}.
}
\examples{
data(parks_sgp)
head(parks_sgp)
}
\keyword{datasets}
