% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_dasymap.R
\name{pop_dasymap}
\alias{pop_dasymap}
\title{Perform dasymetric mapping of (rasterised) population data for a single time period}
\usage{
pop_dasymap(
  pop_polygons,
  pop_perblock_count,
  pop_perblock_density,
  land_relative_density,
  filename = NULL,
  dir_rastertemplate = NULL,
  overwrite = TRUE,
  ...
)
}
\arguments{
\item{pop_polygons}{\code{sf} polygons of the population data for a single time period.
May be extracted from the \code{pop_polygons} list element generated by \code{rasterise_pop()}.}

\item{pop_perblock_count}{Raster of population count per census block for a single time period.
May be extracted from the \code{pop_count_rasters} list element generated by \code{rasterise_pop()}.}

\item{pop_perblock_density}{Raster of population density per census block for a single time period.
May be extracted from the \code{pop_density_rasters} list element generated by \code{rasterise_pop()}.}

\item{land_relative_density}{Raster of the relative density (e.g. suitability, habitability) of land for a single time period.
May be extracted from results generated by \code{rasterise_buildings()} or \code{rasterise_landuse()}.}

\item{filename}{character (optional). Export output raster to disk.}

\item{dir_rastertemplate}{character. Filepath to the raster used to define the pixel resolution, extent, nrow, ncol of
the output raster; object is passed to the '\code{y}' argument in \code{terra::rasterize()}.
Defaults to the template raster generated by the function \code{rasterise_pop()} within the \code{tempdir()}.}

\item{overwrite}{logical. Argument passed to \code{terra::writeRaster()}. If \code{TRUE}, \code{filename} is overwritten.}

\item{...}{Other arguments passed to \code{terra::writeRaster()}.}
}
\value{
Raster of population density, based on relative density values of the land defined in \code{land_relative_density}.
}
\description{
Distribute population data per census block across inhabitable land as defined by \code{land_relative_density}.
Data provided should be a snapshot of the population for a single time period (e.g. one year).
All input data should have a similar projected coordinate reference system
specific to the target area.
}
\examples{
\dontrun{
population_sgp <- data(singapore)
landuse <- data(landuse_sgp)
buildings <- data(buildings_sgp)

# first, rasterise population, landuse & buildings
pop_rasters <- rasterise_pop(population_sgp)

landuse_rasters <- rasterise_landuse(landuse,
                                     land_use = "lu_desc",
                                     subset = c("1" = "RESIDENTIAL",
                                                "2" = "COMMERCIAL & RESIDENTIAL",
                                                "3" = "RESIDENTIAL WITH COMMERCIAL AT 1ST STOREY",
                                                "4" = "RESIDENTIAL / INSTITUTION"),
                                     sf_pop = population_sgp,
                                     match_landuse_pop = "recent")

buildings_rasters <- rasterise_buildings(buildings,
                                     proxy_pop_density = "levels",
                                     year = "year",
                                     sf_pop = population_sgp,
                                     sf_landuse = landuse,
                                     match_buildings_pop = "closest")

# then perform dasymetric mapping on selected (first) year
popdens_raster <- pop_dasymap(pop_polygons = pop_rasters$pop_polygons[[1]],
                              pop_perblock_count = pop_rasters$pop_count[[1]],
                              pop_perblock_density = pop_rasters$pop_density[[1]],
                              land_relative_density = buildings_rasters[[1]],
                              filename = "buildings_popdensity.tif",
                              wopt = list(gdal=c("COMPRESS=LZW")))
}

}
