% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_density_polygonise.R
\name{pop_density_polygonise}
\alias{pop_density_polygonise}
\title{Polygonise population density raster to population count per (building) polygon}
\usage{
pop_density_polygonise(
  input_raster,
  write = TRUE,
  dsn,
  driver = "GeoJSON",
  overwrite = TRUE,
  delete_dsn = TRUE,
  ...
)
}
\arguments{
\item{input_raster}{Population density raster as a SpatRaster object (\code{terra::rast()}).
Output from \code{pop_dasymap()} may be used.}

\item{write}{Whether or not to export the output. Defaults to \code{TRUE}.}

\item{dsn}{Argument passed to \code{sf::st_write()}.}

\item{driver}{character. Argument passed to \code{sf::st_write()}. Defaults to \code{'GeoJSON'}.}

\item{overwrite}{logical. Argument passed to \code{sf::st_write()}. Defaults to \code{TRUE}.}

\item{delete_dsn}{Argument passed to \code{sf::st_write()}. Defaults to \code{TRUE}.}

\item{...}{Other arguments passed to \code{sf::st_write()}.}
}
\value{
Building polygons (sf) with column \code{popcount}.
}
\description{
Convert raster of population density (e.g. from \code{pop_dasymap()} output) to population count per polygon.
Adjacent polygons with similar pixel values are merged. Thus, if input/output rasters
supplied to \code{pop_dasymap()} were at the spatial resolution of individual buildings,
conversion to polygons using this function would result in an output of population count per building.
Input data should have a projected coordinate reference system specific to the target area.
}
